requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
ipympl>=0.9.0  # For interactive plots in Jupyter
jupyter_widgets>=1.1.0

# Data processing
scipy>=1.7.0
xarray>=2022.3.0  # For netCDF handling

# GPT Interface
openai>=1.0.0
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-cov>=2.12.0

# Development tools
black>=22.3.0  # Code formatting
pylint>=2.12.0  # Code analysis
jupyter>=1.0.0  # Notebook support

# Optional dependencies
netCDF4>=1.5.7  # For CTD data
gsw>=3.4.0  # For seawater calculations